Modified files to make links work correctly

+++ Qdel.php
@@ -1,7 +1,4 @@
 <html>
-
-
-<html>
 <head>
     <title>QBank: Question Bank generation made easy.</title>
   
@@ -31,18 +28,66 @@ echo $fileName;
 echo "</i>";
 echo "  has been deleted</h2>";
 
+$type = $_POST['file_type'];
+
 /*Intermediate file location*/
-$myFile = '/home/algoviz-beta/QBank/QBank/OpenDSA/Intermediate_files/'.$fileName.'.txt';
+if($type == 'par')
+{
+$myFile = './Intermediate_files/Parameter/'.$fileName.'.txt';
 unlink($myFile);
 
 deletefromKA();
 
+}
+
+else if($type == 'grp')
+
+{
+$myFile = './Intermediate_files/Group/'.$fileName.'.txt';
+unlink($myFile);
+
+deletefromKA();
+
+}
+
+else
+{
+
+$myFile = './Intermediate_files/Simple/'.$fileName.'.txt';
+unlink($myFile);
+
+deletefromKA();
+
+}
+
+
+
 /*Function that deletes an intermediate file and KA file*/
 
 function deletefromKA() {
+
+
 $fileName = $_POST['file_name'];
-$KAFile = '/home/algoviz-beta/QBank/QBank/OpenDSA/QBank-exercises/'.$fileName.'KA.html';
+
+$type = $_POST['file_type'];
+
+
+if($type == 'par'){
+$KAFile = './Exercises/'.$fileName.'KA.html';
 unlink($KAFile);
+}
+
+
+else if($type == 'grp'){
+$KAFile = './Exercises/'.$fileName.'KA.html';
+unlink($KAFile);
+}
+
+else {
+$KAFile = './Exercises/'.$fileName.'KA.html';
+unlink($KAFile);
+}
+
 }
 
 ?>
@@ -51,11 +96,60 @@ unlink($KAFile);
 
 
 //Displays list of existing questions on screen.
-$pathQ = '/home/algoviz-beta/QBank/QBank/OpenDSA/Intermediate_files';
+$pathQ = './Intermediate_files/Parameter';
 $execCmd = 'ls -rt '.$pathQ.' | tr " " "\t"';
 $output = shell_exec($execCmd);
 $filearray = explode(".txt", $output);
 echo "<table border=1 align=center CELLSPACING=1  CELLPADDING=2 width=20%>";
+echo "<h3>";
+echo "Parameterised Q";
+echo "</h3>";
+
+
+
+for($i=0; $i<sizeof($filearray); $i++){
+	
+	echo "<tr>";
+	echo "<td>";
+	echo "<p>";
+	echo $filearray[$i]; 
+	echo "<p>";
+	echo "</td>";
+	echo "</tr>";
+}
+echo "</table>";
+
+$pathQ = './Intermediate_files/Group';
+$execCmd = 'ls -rt '.$pathQ.' | tr " " "\t"';
+$output = shell_exec($execCmd);
+$filearray = explode(".txt", $output);
+echo "<table border=1 align=center CELLSPACING=1  CELLPADDING=2 width=20%>";
+
+echo "<h3>";
+echo "Group Q";
+echo "</h3>";
+
+for($i=0; $i<sizeof($filearray); $i++){
+	
+	echo "<tr>";
+	echo "<td>";
+	echo "<p>";
+	echo $filearray[$i]; 
+	echo "<p>";
+	echo "</td>";
+	echo "</tr>";
+}
+echo "</table>";
+
+$pathQ = './Intermediate_files/Simple';
+$execCmd = 'ls -rt '.$pathQ.' | tr " " "\t"';
+$output = shell_exec($execCmd);
+$filearray = explode(".txt", $output);
+echo "<table border=1 align=center CELLSPACING=1  CELLPADDING=2 width=20%>";
+
+echo "<h3>";
+echo "Simple Q";
+echo "</h3>";
 for($i=0; $i<sizeof($filearray); $i++){
 	
 	echo "<tr>";
@@ -73,10 +167,6 @@ echo "</table>";
          <a href="index.html">About QBank</a> | 
          <a href="index.html">Contact Us</a>
       </div>
-
-
-
-
 
 </body>
 </html>
